+++ src/analyze/src/components/settings/Database.vue
@@ -65,7 +65,7 @@
           </div>
           <div class="control">
             <div class="select">
-              <select v-model="connectionDialect" placeholder="dialect">
+              <select v-model="connectionDialect">
                 <option value="" disabled selected>Dialect</option>
                 <option value="postgresql">PostgreSQL</option>
                 <option value="mysql">MySQL</option>
@@ -146,6 +146,7 @@
   </div>
 </template>
 <script>
+import store from '@/store';
 import { mapState, mapGetters } from 'vuex';
 
 export default {
@@ -165,8 +166,12 @@ export default {
     };
   },
 
-  mounted() {
-    this.$store.dispatch('settings/getSettings');
+  beforeRouteEnter(to, from, next) {
+    store.dispatch('settings/getSettings')
+      .then(next)
+      .catch(() => {
+        next(from.path);
+      });
   },
 
   computed: {