+++ src/analyze/src/router/index.js
@@ -1,9 +1,6 @@
 import Vue from 'vue';
 import Router from 'vue-router';
 import Repo from '@/views/Repo';
-import Projects from '@/views/Projects';
-import store from '@/store/';
-import Start from '@/views/Start';
 import Design from '@/views/Design';
 import Dashboards from '@/views/Dashboards';
 import NotFound from '@/views/NotFound';
@@ -23,55 +20,42 @@ const router = new Router({
     },
     {
       path: '/',
-      name: 'home',
-      redirect: '/projects',
+      redirect: '/files',
     },
     {
-      path: '/projects/',
-      name: 'projects',
-      component: Projects,
-    },
-    {
-      path: '/start/',
-      name: 'start',
-      component: Start,
-    },
-    {
-      path: '/projects/:projectSlug/files/',
-      name: 'projectFiles',
+      path: '/files/',
+      name: 'Repo',
       component: Repo,
     },
     {
-      path: '/projects/:projectSlug/analyze/:model/:design',
-      name: 'analyze',
+      path: '/analyze/:model/:design',
+      name: '',
       component: Design,
     },
     {
-      path: '/projects/:projectSlug/analyze/:model/:design/reports/report/:slug',
-      name: 'design_report',
+      path: '/analyze/:model/:design/reports/report/:slug',
+      name: 'Report',
       component: Design,
     },
     {
-      path: '/projects/:projectSlug/dashboards/',
-      name: 'dashboards',
+      path: '/dashboards/',
+      name: 'Dashboards',
       component: Dashboards,
     },
     {
-      path: '/projects/:projectSlug/dashboards/dashboard/:slug',
-      name: 'dashboard',
+      path: '/dashboards/dashboard/:slug',
+      name: 'Dashboard',
       component: Dashboards,
     },
     {
-      path: '/projects/:projectSlug/settings',
-      name: 'settings',
+      path: '/settings',
+      name: 'Settings',
       component: Settings,
       children: [{
         path: 'roles',
-        name: 'roles',
         component: SettingsRoles,
       }, {
         path: 'database',
-        name: 'database',
         component: SettingsDatabase,
       }],
     },