+++ src/meltano/api/controllers/repos.py
@@ -2,14 +2,11 @@ import json
 import os
 from os.path import join
 from pathlib import Path
-from functools import wraps
 
 import markdown
 from flask import Blueprint, jsonify, request
 
-from .project_helper import project_api_route, project_from_slug
-
-from meltano.core.project import Project, ProjectNotFound
+from meltano.core.project import Project
 from meltano.core.utils import decode_file_path_from_id
 from meltano.core.compiler.project_compiler import ProjectCompiler
 from meltano.api.security import api_auth_required
@@ -67,8 +64,8 @@ def before_request():
 
 
 @reposBP.route("/", methods=["GET"])
-@project_from_slug
-def index(project):
+def index():
+    project = Project.find()
     onlyfiles = [f for f in project.model_dir().iterdir() if f.is_file()]
 
     path = project.model_dir()
@@ -85,7 +82,7 @@ def index(project):
         "reports": {"label": "Reports", "items": reportsFiles},
         "tables": {"label": "Tables", "items": []},
     }
-    onlydocs = path.parent.glob("*.md")
+    onlydocs = project.model_dir().parent.glob("*.md")
     for d in onlydocs:
         file_dict = MeltanoAnalysisFileParser.fill_base_m5o_dict(d, str(d.name))
         sortedM5oFiles["documents"]["items"].append(file_dict)
@@ -124,11 +121,11 @@ def index(project):
 
 
 @reposBP.route("/file/<unique_id>", methods=["GET"])
-@project_from_slug
-def file(unique_id, project):
+def file(unique_id):
     file_path = decode_file_path_from_id(unique_id)
     (filename, ext) = os.path.splitext(file_path)
     is_markdown = False
+    project = Project.find()
     path_to_file = project.model_dir(file_path).resolve()
     with open(path_to_file, "r") as read_file:
         data = read_file.read()
@@ -145,7 +142,8 @@ def file(unique_id, project):
         )
 
 
-def lint_all(compile, project):
+def lint_all(compile):
+    project = Project.find()
     compiler = ProjectCompiler(project)
     try:
         compiler.parse()
@@ -174,39 +172,23 @@ def handle_file_not_found(e):
 
 
 @reposBP.route("/lint", methods=["GET"])
-@project_from_slug
-def lint(project):
-    return lint_all(False, project_slug)
+def lint():
+    return lint_all(False)
 
 
 @reposBP.route("/sync", methods=["GET"])
-@project_from_slug
-def sync(project):
-    return lint_all(True, project)
-
-
-@reposBP.route("/test", methods=["GET"])
-def db_test():
-    design = Design.query.first()
-    return jsonify({"design": {"name": design.name, "settings": design.settings}})
+def sync():
+    return lint_all(True)
 
 
 @reposBP.route("/models", methods=["GET"])
-@project_from_slug
-def models(project):
+def models():
+    project = Project.find()
     topics = project.root_dir("model", "topics.index.m5oc")
     topics = json.load(open(topics, "r"))
     topics = next(M5ocFilter().filter("view:topic", [topics]))
+
     return jsonify(topics)
-
-
-@reposBP.route("/designs/", methods=["GET"])
-def designs():
-    designs = Design.query.all()
-    designs_json = []
-    for design in designs:
-        designs_json.append(design.serializable())
-    return jsonify(designs_json)
 
 
 @reposBP.route("/tables/<table_name>", methods=["GET"])
@@ -219,9 +201,10 @@ def table_read(table_name):
 
 
 @reposBP.route("/designs/<topic_name>/<design_name>", methods=["GET"])
-@project_from_slug
-def design_read(topic_name, design_name, project):
+def design_read(topic_name, design_name):
     permit("view:design", design_name)
+
+    project = Project.find()
     topic = project.root_dir("model", f"{topic_name}.topic.m5oc")
     with topic.open() as f:
         topic = json.load(f)